requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.15.1
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
openai-whisper==20231117
PyAudio==0.2.14
pydub==0.25.1
regex==2024.5.15
requests==2.32.3
SpeechRecognition==3.10.4
sympy==1.12.1
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
